Add special case for nvme device names.
Partitions follow `nvme_n_p_`, where the `p` is between the device name
and the partition number. Other types of storage devices do not add the
`p` between the two, so if blocks are needed to distinguish between the
two. A simple regex to store as boolean variable will suffice.