add rust language to dev list
Debian does not have rustup in the repos but instead packages several
crates in its repos. Install these instead, and change default cargo
behavior to use system crates locally and ignore the global (online)
index.

This is to be undone should Debian ever package rustup itself. The crate
packages are present because of its internal packaging policy that each
deb binary must also provide its sources. It's not intended as a mirror
for crates.io and will likely have out-of-date and missing packages.